Cleaning up: Removed "explore the data" + tooltips, renumbered sections
Marti did not think the data was very interesting to explore, so we have
removed the interactive version. This was previously container-2 and included
code for tooltips as well as the raw CSV files.

We have renumbered sections accordingly, notably making the time trends into
container-2 and starting numbering at 1 instead of 0. There was also an error
in chart 4 where we had been referencing the dimensions of graphic 2 instaed of
4 -- this has been corrected.

Also, we removed the video file originally shown at the top since Marti thought
it was distracting.