Graph 2: re-added tooltips + points (but the points have opacity 0)
Since the text is no longer on top of the graph, the user can interact with
the graph as they view it. A tooltip div will show specific percentage changes
if any of the "points" are hovered over. Points are not actually visible since
the opacity is set to 0 -- this was done because they otherwise showed up
before the animation was finished. The radius for the points was also increased
so the user is more likely to hover over the point. If they happen to hover
over the line, they will just see the "title" element